requirements.txt
flask~=2.3.3
python-dotenv~=1.0.0
flask-login
flask-bcrypt
flask-migrate
pdf2image~=1.16.3
flask-mail
pypdf~=3.17.2
openai~=1.3.8
SQLAlchemy~=2.0.21
requests~=2.31.0
Faker~=19.6.2
Werkzeug~=2.3.7
alembic~=1.13.0